Move forecast components to subfolder in /components
