+++ src/components/forecast/Forecast.js
@@ -3,7 +3,7 @@ import React from 'react';
 import './Forecast.css';
 import Day from './Day';
 import DayView from './DayView';
-import { dtToDate } from './helpers';
+import { dtToDate } from '../../helpers';
 
 const Forecast = ({ city }) => {
   const [forecastError, setForecastError] = useState(null);
